Reduce devtools_test.dart flakiness (#1086)
The `can not launch devtools for the same app in multiple tabs` test appears to be flaky because we attempt to launch DevTools before the page is fully loaded. There isn't a good asynchronous signal for this so as an alternative we wait for the page content to say `Hello World!`. This  closes https://github.com/dart-lang/webdev/issues/1084.