Remove frame dependency on ScriptID (#1093)
Sending all `scriptParsedEvent`s is on the critical path for DevTools IPL. These events are used to blackbox the SDK and translate paused frames to Dart. With this change, we no longer need all events, just the `scriptParsedEvent` for the Dart SDK. I'll follow this change by modifying the Dart Debug Extension to only send that single event.