Dart Debug Extension Authentication (#1277)
When a user clicks the Dart Debug Extension they may not have the required auth cookies which will cause unexpected behavior. Now detect if a connection can not be established, and prompt the user for authentication.

Ideally we wouldn't require the user to reinitiate the debug workflow but this is a good short term solution that will solve a number of issues.