Improve Dart Debug Extension workflow (#1309)
As a part of the injected metadata we now also post an event to the page. This event is consumed by the Dart Debug Extension. This allows us to conditionally enable the extension which greatly improves usability. Users should no longer get an alert window indicating that the selected tab does not contain a Dart application.

Before we publish the Dart Debug Extension we need to make sure the `package:dwds` logic is available in all tools, e.g. Webdev.

Some notable changes:
- The various `chrome.*` listeners are now all top level singletons
  - Previously we were adding tons of listeners and never cleaning them up (not good)
- We needed to resize the icons, otherwise Chrome fails when dynamically setting the icon
- The Dart Debug Extension now injects a content script to forward the Dart ready event to extension backend
- We no longer do a timeout while collecting execution contexts
  - Since we know the tab has a running Dart application we can properly wait for the correct context (this reduces flakiness)
 - No longer give an alert when you detach the debugger
   - I found this annoying