Avoid OOM errors when logging many events
Revert to attaching the IsolateRef instead of the Isolate to the postEvent
and registerExtension events. The Isolate is much larger and was causing
systems to run out of memory when event traffic was high like in Flutter
where they are sent on every animation frame.

See: https://github.com/flutter/flutter/issues/87100