Fix hot restart and breakpoints races
Multiple breakpoints restored ad the isolate start can cause
the app to crash in chrome, serialize them using a mutex to
prevent the crash.

Hot restart runs main immediately, followed by createIsolate
that restores breakpoints, sometimes too late, causing non-
removable breakpoints and failures in the UI. Make the dwds
run main after createIsolate is done, to make sure breakpoints
are all set before the run.

Hot restart tries to resume the exiting isolate, and checks if
it resumed before that, which causes various races or chrome
requiring to be in focus, depending on the way the check is done.
Remove the check and just catch exception in case the isolate
is already running.