Support multiple DDC entrypoints on the same page
Currently, if webdev is serving a page that loads more than one dart
entrypoint compiled with DDC, the `main` that gets hoisted to the global
scope along with the extra code used by webdev will run once for each
entrypoint resulting in already-defined errors.

This commit works around that issue by updating the code that lives in
the global scope to check for the existence of its vars before defining
them. Additionally, it will augment those vars so that each entrypoint
can contribute instead of only the first.

Additionally, the digests handler has been updated to accept multiple
values for the `entrypoint` query parameter. It will return the digests
for each entrypoint combined into a single response.

This commit stops short of fixing debugging for multiple entrypoints on
the same page; only the first entrypoint will be debuggable. It is possible
to make each entrypoint debuggable, but additional work is required and
the approach will likely need much more feedback, so I've left it out.