Webdev serve: add an option to pass user data directory to chrome (#1491)
* Webdev serve: add an option to pass user data directory to chrome

Webdev launches chrome using temp profile by default, so the users
cannot use their extensions while debugging. Allow launching chrome
in a window that is signed in as the user's default profile.

Notes:

There seems to be a limitation in chrome that only uses the existing
session with --user-data-dir. The existing session cannot open a
debug port that dwds needs.

To work around this issue, we copy the default profile to a temp
directory, and pass it to chrome. That seems to work, but takes time
to copy the profile. Issue a performance warning in that case.

Note that subsequent updates will only update modified files, so the
performance hit becomes smaller over time.

Closes: https://github.com/dart-lang/webdev/issues/1490

* Updated changelog and test

* Fix some test failures

* Fixed windows test faiures

* Fix test faillures on linux

* Fixed format

* Fix failing tests on Windows

* Add automadic detection of Chrome user data directory

* Fixed failing test on Windows

* Extend timeout for chrome test to prevent flakes

* Added links to issue for supporting the feature on windows