Add ignores for deprecated web libraries with a TODO and move dwds to 24.4.0-wip (#2559)
Dart web libraries are now deprecated. This triggers a lint which we fail on. We should migrate these to dart:js_interop and package:web, but for now, ignore to make CI happy.

This also moves the CHANGELOG, pubspec, and version.dart files to 24.4.0-wip, which is needed for proper_release_test to pass.