[dwds] Split hot reload callback into 2, disable breakpoints in changed files, and publish 24.3.10
https://github.com/dart-lang/sdk/issues/60186

Breakpoints are currently broken when doing a hot reload. They either
are not re-mapped or point to the wrong location. For now, we should
disable them in the changed libraries. In order to do this, we split
the hot reload implementation into 2 in order to get the list of changed
libraries. The call to hot reload will later use the cached list to
call out to the embedder.