Support breakpoints for hot reload (#2627)
Adds support for properly reregistering breakpoints across
a hot reload. Like with hot restart, leverages the kPausePostRequest
event in order to reregister breakpoints. DWDS does the initial
work to load the new sources in ("hotReloadStart"), and pauses and
sends the request. Upon disabling existing breakpoint, reregistering
and registering new breakpoints and resuming, DWDS will resume the
app and push the new sources into the runtime ("hotReloadEnd").

Also reinitializes some cached information in order to compute new
sources. Note that it's very conservative for now. A future change
should optimize the invalidation.

See #2628 for that work.

Adds tests for the following cases:

Editing a line with a breakpoint.
Adding a line before a breakpoint.
Removing a line before a breakpoint.
Adding a new file with a breakpoint in it.
Breakpoint in previously captured code is removed.
There's no pause when pause_isolates_on_start is false.
dart-lang/sdk#60186

Includes a light refactor of testing to create temp directories wherever the TestProject needs to be edited. TestProjects should specify whether it's meant to be edited via editable and list its dependencies so they can be copied too.