Address flakes from using temp directory to run tests
https://github.com/dart-lang/webdev/commit/2172ba7024d8885afc30f6634d98c00a003bb0cc
added support to run tests in a temporary directory.

This results in two flaky issues:

1. On Windows, build_daemon tests may fail to delete the
temp directory because the process may have not been torn
down yet, so it may still be accessing the file system.
There was an initial retry after 1 second, but that appears
to be not enough looking at a recent test run.
2. If a test times out, its tearDown may not be called. In
this case, the ResidentWebRunner in frontend_server may not
restore the current directory in the LocalFileSystem. This
leads to cascading failures in subsequent tests due to no
longer being in a path that contains 'webdev'. See
https://github.com/dart-lang/webdev/actions/runs/15989286213/job/45099373212?pr=2641
for an example. See https://github.com/dart-lang/test/issues/897
as well for tracking work to call tearDown on timeouts.

To address the above issues:

1. Increase the delay between the two tries and assert this
only occurs on Windows.
2. Cache the current directory so that it can be used in
utilities.dart with the same (correct) value every time.