[dwds] Wait for scripts to be parsed on a hot restart and publish DWDS 25.0.0
Like https://github.com/dart-lang/webdev/issues/2640, we
should wait until scripts are parsed before continuing a hot restart.
Otherwise, metadata can be stale and breakpoints may be placed in the
wrong files.

For now, we only do this in the DDC library bundle format as we don't
have a way to fetch the changed libraries in the AMD format.

- hotReloadSourcesUri is repurposed to be reloadedSourcesUri, which is
a breaking change. This file is now used for both hot restart and hot
reload to detail the changed files across either.
- Injected client is changed to return the mapping within this uri when
a hot reload is executed.
- Completer is added to ChromeProxyService to wait until all scripts are
parsed before recreating the isolate, which will reinitialize all metadata.
- Fix a preexisting race condition where we don't listen for a kIsolateStart
event early enough.

This should help fix the Windows flakes for hot_restart_breakpoints_test.dart.