Wait a second before making edits to files on Windows
dart:io does not have millisecond-level precision on file stats
and instead, the modified time is rounded to the nearest second.
If edits and recompiles are done within the same second as the
previous compile, frontend_server_common can't tell that the files
have changed. Instead, wait a second before editing. In order to
support this, edits are staged at once. Without this, we'd be
waiting multiple seconds for several edits.