Use evaluate instead of periodic timer to test dot shorthands (#2687)
Fixes https://github.com/dart-lang/webdev/issues/2638

Fields are cached after their first execution. When debugging, this
means we may not hit the body of the field when using the AMD
module format. The DDC library bundle format always has a getter for
fields, even with the cached value, however.

To make the test consistent, we should avoid any previous executions
of the field by using an evaluate call instead of a periodic timer
in main.

This requires moving the dot shorthands test to its own fixture.