Fix awaiting for  future
In https://github.com/dart-lang/webdev/pull/2491, the reloader was changed to
pass two arguments instead of an object with two fields into dartLibrary.reload.
As a result, the javascript side couldn't find the promise in a field of the
first argument and ignored the second argument. Usually it is fine, but this
broke postponing main launch until the attached debugger can configure breakpoints.

Again, this is not a big deal for interactive long-running apps, but if the main is
fast and short (e.g. tests), this effectively prevents interactive debugging from an IDE.