[ DWDS ] Reestablish WS connections with app and Chrome debugger
When the computer enters sleep mode, Chrome does not attempt to
reestablish websocket connections when the computer wakes up. This would
break debugging of web applications when DWDS is configured to
communicate with the Chrome debugger or the debugged application is
using web sockets to communicate with DWDS.

This change adds some retry logic to reestablish these web socket
connections from the injected client and to the Chrome debugger.

Related to https://github.com/flutter/flutter/issues/179886